Using HOCs in exports of ArtistQuestionScreen and GenreQuestionScreen
